src/backend/youtube_scrapper_for_internship_main/requirements.txt
# Flask backend
flask
flask-cors
flask-sqlalchemy
qrcode
pillow

# YouTube scraper
google-api-python-client==2.121.0
python-dotenv

flask
flask-cors
celery[redis]

psycopg2-binary
sqlalchemy